[core] Move AuxClasspathLoader from java to core
