+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symbols/ClassLoadingChildFirstTest.java
@@ -16,10 +16,10 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import net.sourceforge.pmd.internal.util.AuxClasspathLoader;
 import net.sourceforge.pmd.internal.util.ClasspathClassLoader;
 import net.sourceforge.pmd.lang.JvmLanguagePropertyBundle;
 import net.sourceforge.pmd.lang.java.JavaLanguageModule;
-import net.sourceforge.pmd.lang.java.internal.AuxClasspathLoader;
 import net.sourceforge.pmd.lang.java.internal.JavaLanguageProperties;
 import net.sourceforge.pmd.lang.java.types.JClassType;
 import net.sourceforge.pmd.lang.java.types.TypeSystem;
@@ -53,7 +53,7 @@ class ClassLoadingChildFirstTest {
         languageProperties.setProperty(JvmLanguagePropertyBundle.AUX_CLASSPATH, classpath);
 
         try (AuxClasspathLoader auxClasspathLoader = AuxClasspathLoader.create(classpath, false)) {
-            TypeSystem typeSystem = TypeSystem.usingClasspath(auxClasspathLoader);
+            TypeSystem typeSystem = TypeSystem.usingClasspath(name -> auxClasspathLoader.findResource(name));
             JClassType voidClass = typeSystem.BOXED_VOID;
             List<JMethodSymbol> declaredMethods = voidClass.getSymbol().getDeclaredMethods();
             assertThat(declaredMethods, hasSize(1));
